Правки по замечаниям
