+++ index.php
@@ -1,8 +1,8 @@
 <?php
 require_once __DIR__ . '/vendor/autoload.php';
-use config\AvailableActions;
+use components\Task;
 
-$myTask = new AvailableActions([
+$myTask = new Task([
 	'idClient' => 5,
 	'idExecutor' => 3,
 ]);
